Ignore "tags" files only in the root
It appears this was added to due "ctags" which is an code indexing tool.
It appears to put its data in a "tags" file in the root of the project
by default.

We want to still be able to use "tags" files and dirs inside our
projects, for instance /app/views/tags/.